+++ http.rest
@@ -17,7 +17,7 @@ Content-Type: application/json
 
 {
   "email": "Jill@Crown.lt",
-  "password": ""
+  "password": "secret1234"
 }
 
 ###
