+++ src/middleware.js
@@ -12,7 +12,9 @@ function showBody(req, res, next) {
 async function validateUser(req, res, next) {
   // validuoti gauta email ir password
   const schema = Joi.object({
+    // eslint-disable-next-line newline-per-chained-call
     email: Joi.string().trim().email().lowercase().required(),
+    // eslint-disable-next-line newline-per-chained-call
     password: Joi.string().trim().min(5).max(10).required(),
   });
